==> notebooks/01_preprocesamiento_NER_POS.ipynb <==
Pedido: Obtener NER y POS sobre el mismo dataset sobre todas las conversaciones etiquetadas manualmente.

Los datasets son:

- ecu_50_llamadas_sample_annotated.xlsx
    Primera muestra de 50 llamadas etiquetadas por el pasante mejia.

- ecu_50_llamadas_sample_annotated_v2.xlsx
    Segunda muestra de 50 llamadas etiquetada en conjunto por mi y por el pasante mejia.

- ecu_150_llamadas_sample_annotated_v3.xlsx
    Tercera muestra pero esta vez de 150 llamadas etiquetadas por Andres Sebastian Guerrero Zeas

En total se tienen 250 llamadas etiquetadas con acto de diálogo.

!which python

import os

os.environ["CUDA_VISIBLE_DEVICES"] = "2"

Spacy tiene opcion para NER pero no con un modelo
Transformer sino con "es_core_news_lg".
El modelo transformer "es_dep_news_trf" puede hacer POS

# !python -m spacy download es_core_news_lg

import spacy
spacy.require_gpu()

# Ver los componentes del modelo transformer
nlp = spacy.load("es_dep_news_trf")
print(nlp.pipe_names)

# obtener NER
nlp = spacy.load("es_core_news_lg")
doc = nlp("La compañía UDA está lanzando un nuevo libro llamado 'Aprendiendo' del autor Santiago")

for ent in doc.ents:
    print(ent.text, ent.label_)

# Obtener POS
nlp = spacy.load("es_dep_news_trf")

doc = nlp("La empresa UDA publicó un libro nuevo.")

for token in doc:
    print(token.text, token.pos_, token.dep_)


import pandas as pd
from tqdm import tqdm
tqdm.pandas()

df = pd.read_excel("../datasets/ecu_250_sample_annotated_union_v1v2v3.xlsx")

df.head(5)

from sklearn.feature_extraction.text import TfidfVectorizer
import spacy

spacy.require_gpu()

nlp_ner = spacy.load("es_core_news_lg")
nlp_pos = spacy.load("es_dep_news_trf")

def extract_ner(text):
    if not isinstance(text, str) or text.strip() == "":
        return []
    doc = nlp_ner(text)
    return [(ent.text, ent.label_) for ent in doc.ents]

def extract_pos(text):
    if not isinstance(text, str) or text.strip() == "":
        return []
    doc = nlp_pos(text)
    return [(token.text, token.pos_) for token in doc]

df["NER"] = df["ATA_TEXTO_TAGGED"].progress_apply(extract_ner)
df["POS"] = df["ATA_TEXTO_TAGGED"].progress_apply(extract_pos)

# stop words desde spacy

#stopwords_es = nlp_ner.Defaults.stop_words

# TF-IDF
# Devuelve una matriz (sparse matrix) de forma:
# (n_documentos, n_features)
# n_documentos = filas dataset
# n_features palabras más frecuentes, definidas por max_features
# vectorizer = TfidfVectorizer(
#     lowercase=True,
#     stop_words=list(stopwords_es),
#     ngram_range=(1,2),
#     max_features=3000, # Número máximo de columnas (palabras) a conservar.
#     # min_df=5      # ignora palabras que aparecen en menos de 5 documentos
#     # max_df=0.8    # ignora palabras que aparecen en >80% de los documentos
# )

#tfidf_matrix = vectorizer.fit_transform(df["ATA_TEXTO_TAGGED"].fillna(""))


df

a = df.iloc[0]["POS"]

for e in a:
    print(e)
    print(type(e))

# df.to_csv("250_conversaciones_ner_pos.csv", index=False)
# df.to_excel("250_conversaciones_ner_pos.xlsx", index=False)
